// backend/src/modules/category/model/dto.ts
import z = require("zod");
import { ModelStatus } from "../../../share/model/base-model";



export const CategoryCreateSchema = z.object({
    name: z.string().min(3, "Name is required and cannot be empty."),
    image: z.string().nullable().optional(),
    description: z.string().nullable().optional(),
    position: z.number().min(0).default(0),
    parentId: z.string().uuid().nullable().optional(),
});

export type CategoryCreateDTO = z.infer<typeof CategoryCreateSchema>;


export const CategoryUpdateSchema = z.object({
    name: z.string().min(2, "Name is required and cannot be empty.").optional(),
    image: z.string().nullable().optional(),
    description: z.string().max(255, "Description cannot exceed 255 characters.").nullable().optional(),
    position: z.number().min(0).optional(),
    parentId: z.string().uuid().nullable().optional(),
    status: z.nativeEnum(ModelStatus).optional(),
})

export type CategoryUpdateDTO = z.infer<typeof CategoryUpdateSchema>;

export const CategoryCondDTOSchema = z.object({
    name: z.string().min(2, "Name is required and cannot be empty.").optional(),
    parentId: z.string().uuid().optional(),
    status: z.nativeEnum(ModelStatus).optional(),
})

export type CategoryCondDTO = z.infer<typeof CategoryCondDTOSchema>;